Conflicts:
	lime/Assets.hx
	lime/audio/AudioBuffer.hx
	lime/graphics/Image.hx
	lime/graphics/cairo/CairoFont.hx
	lime/graphics/cairo/CairoFontOptions.hx
	lime/graphics/cairo/CairoImageSurface.hx
	lime/system/JNI.hx
	lime/system/System.hx
	lime/system/ThreadPool.hx
	lime/utils/ByteArray.hx
	project/lib/sdl
	project/src/ExternalInterface.cpp
	project/src/graphics/cairo/CairoBindings.cpp

+++ project/src/video/backend/wmf/WMFVideo.cpp
@@ -136,7 +136,7 @@ HRESULT WMFVideo::QueryInterface(REFIID riid, void** ppv)
     static const QITAB qit[] = 
     {
         QITABENT(WMFVideo, IMFAsyncCallback),
-        { 0 }
+        { (int)0 }
     };
     return QISearch(this, qit, riid, ppv);
 }
